db(appointment): change properity pateint and doctor ids

+++ database/migrations/2024_07_03_113618_create_appointments_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('appointments', function (Blueprint $table) {
             $table->id();
-            $table->integer('doctor_id');
-            $table->integer('patient_id');
+            $table->unsignedBigInteger('doctor_id');
+            $table->unsignedBigInteger('patient_id');
             $table->date('date');
             $table->time('time');
             $table->string('reason');
